Internal

Functions

GeneralizedTransferMatrixMethod.eigen_trans_ref — Function

eigen_trans_ref(Δ)

Return the four eigenvalues and vectors of the Δ matrix sorted by propagation direction.

The first two entries represent the forward propagating (transmitted) waves, the last two entries the backward propagating (reflected) waves.

Input

Δ – Square matrix.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.eigen_sorted — Function

eigen_sorted(ζ, ϵ, μ, Δ)

Return the sorted eigenvalues, eigenvectors, and Poynting vectors of Δ(ζ, ϵ, μ).

Input

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

Δ – Δ matrix.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.poynting_vector — Function

poynting_vector(Ψ, ζ, ϵ, μ)

Calculate the Poynting vector from the eigenvector Ψ of Δ.

Input

Ψ – Eigenvector of Δ of the form (Ex, Hy, Ey, -Hx)ᵀ.

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.C_p — Function

C_p(F₁, F₂)

Calculate the ordering parameter C from for two quantities F₁ and F₂.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.lorentz_osc — Function

lorentz_osc(f, fₗₒ, fₜₒ, γ)

Calculate a single lorentz oscillator (without ϵ∞).

Input

f – Frequency.

fₗₒ – Frequency of the longitudinal optical phonon.

fₜₒ – Frequency of the transverse optical phonon.

γ – Damping factor of the Lorentzian line shape.

Reference

The definition is taken from Álvarez-Pérez 2020.

GeneralizedTransferMatrixMethod.ϵ_drude — Function

ϵ_drude(ω, ωₚ, γ)

Calculate the permitivity from Drude model.

Input

f – Frequency.

fₚ – Plasma frequency.

γ – Mean collision rate.

GeneralizedTransferMatrixMethod.ϵ_x_MoO₃ — Function

ϵ_x_MoO₃(λ)

Calculate the x principal component of the permitivity tensor of MoO₃.

Input

λ – Wavelength [m].

Reference

The parameters are taken from Álvarez-Pérez 2020.

GeneralizedTransferMatrixMethod.ϵ_y_MoO₃ — Function

ϵ_y_MoO₃(λ)

Calculate the y principal component of the permitivity tensor of MoO₃.

Input

λ – Wavelength [m].

Reference

The parameters are taken from Álvarez-Pérez 2020.

GeneralizedTransferMatrixMethod.ϵ_z_MoO₃ — Function

ϵ_z_MoO₃(λ)

Calculate the z principal component of the permitivity tensor of MoO₃.

Input

λ – Wavelength [m].

Reference

The parameters are taken from Álvarez-Pérez 2020.

Matrices

GeneralizedTransferMatrixMethod.Λ — Constant

Λ

Λ matrix (see reference). Changes the order of the Γ matrix to follow Yeh's formalism.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.euler_mat — Function

euler_mat(θ, ϕ, ψ)

Calculate the Euler matrix from the Euler angles θ, ϕ, and ψ.

    euler_mat(layer::Layer)

Calculate the Euler matrix for a Layer.

GeneralizedTransferMatrixMethod.M_mat — Function

M_mat(ϵ, μ)

Calculate the M matrix (see reference).

Input

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

Notes

Only non-optically active media with isotropic permeability are considered:

\[\bar\mu = \mu\bar1 \quad \bar\rho_1 = \bar\rho_2 = \bar0\]

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.a_mat — Function

a_mat(M, ζ)

Calculate the a matrix (see reference).

Input

M – 6x6 M matrix (see reference).

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.Δ_mat — Function

Δ_mat(ζ, ϵ, μ, M, a)

Calculate the Δ matrix (see reference).

Input

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

M – M matrix.

a – a matrix.

Reference

The definition is taken from Passler and Paarmann 2017.

GeneralizedTransferMatrixMethod.γ_mat — Function

γ_mat(q, ζ, ϵ, μ)

Calculate the normalized γ vectors and return them in matrix form.

Input

q – Sorted eigenvalues of Δ.

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

Notes

The field vectors are given by the rows of the matrix, so the indexing scheme follows the one in the references.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

GeneralizedTransferMatrixMethod.dynamical_mat — Function

dynamical_mat(q, ζ, ϵ, μ γ)

Calculate the dynamical matrix A of an interface. Returns an array of type SMatrix for faster matrix operations.

Input

q – Sorted eigenvalues of Δ.

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

ϵ – 3x3 permitivity tensor in the lab frame.

μ – Scalar permeability.

γ – γ Matrix.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

GeneralizedTransferMatrixMethod.propagation_mat — Function

propagation_mat(q, λ, d)

Calculate the propagation matrix P. Returns an array of type SMatrix for faster matrix operations.

Input

q – Sorted eigenvalues of Δ.

λ – Wavelength [m].

d – Thickness of the material [m].

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).