Internal

The references are displayed at the bottom instead of directly in the docstring because auf Documenter.jl.

GeneralizedTransferMatrixMethod.References — Constant

Dictionary of references used in this Package.

Constants

GeneralizedTransferMatrixMethod.c₀ — Constant

Speed of light in vacuum $[\frac{m}{s}]$.

GeneralizedTransferMatrixMethod.ϵ₀ — Constant

Vacuum permittivity $[\frac{F}{m}]$.

GeneralizedTransferMatrixMethod.μ₀ — Constant

Vacuum permeability $[\frac{N}{A^2}]$.

GeneralizedTransferMatrixMethod.SIdentity — Constant

3D identity matrix using StaticArrays.jl.

TMM matrices

GeneralizedTransferMatrixMethod.J — Constant

All-ones matrix as defined in Eq. 2.50 in [1].

References

GeneralizedTransferMatrixMethod.ν_mats — Function

ν_mats(qₓ, ϵ, μ, ξ, ζ)

Calculate the coefficients of the auxiliary phasors $e_z$ and $h_z$, as described in Eq. 2.46 in [1]. Returns two matrices of the form:

\[\begin{pmatrix} \nu^{ee}_{zx} & & & \\ & \nu^{ee}_{zy} & & \\ & & \nu^{eh}_{zx} & \\ & & & \nu^{eh}_{zy} \end{pmatrix}\]

and

\[\begin{pmatrix} \nu^{he}_{zx} & & & \\ & \nu^{he}_{zy} & & \\ & & \nu^{hh}_{zx} & \\ & & & \nu^{hh}_{zy} \end{pmatrix}.\]

Arguments

qₓ: Normalized in-plane component of the wavevector.

ϵ: Permittivity tensor.

μ: Permeability tensor.

ξ, ζ: Optical rotation tensors.

Examples

julia> GeneralizedTransferMatrixMethod.ν_mats(0.3, 3ones(3,3), ones(3,3), zeros(3,3), zeros(3,3))
([-1.0 0.0 0.0 0.0; 0.0 -1.0 0.0 0.0; 0.0 0.0 0.0 0.0; 0.0 0.0 0.0 -3.3356409519815207e-10], [0.0 0.0 0.0 0.0; 0.0 1.0006922855944562e-9 0.0 0.0; 0.0 0.0 -1.0 0.0; 0.0 0.0 0.0 -1.0])

References

GeneralizedTransferMatrixMethod.P_mat — Function

P_mat(ω, qₓ, ϵ, μ, ξ, ζ)

Calculate the $P$ matrix according to Eq. 2.49 in [1].

Arguments

ω: Angular frequency [1/s] $[\frac{1}{s}]$.

qₓ: Normalized in-plane component of the wavevector.

ϵ: Permittivity tensor.

μ: Permeability tensor.

ξ, ζ: Optical rotation tensors.

Examples

julia> GeneralizedTransferMatrixMethod.P_mat(1.8e14, 0.3, 3ones(3,3), ones(3,3), zeros(3,3), zeros(3,3))
4×4 StaticArraysCore.SMatrix{4, 4, Float64, 16} with indices SOneTo(4)×SOneTo(4):
 -1.80125e5   0.0           0.0         0.0
  0.0        -1.80125e5     0.0         0.0
  0.0         0.000180249  -1.80125e5   0.0
  0.0         0.0           0.0        -1.80125e5

References

GeneralizedTransferMatrixMethod.M_mat — Function

M_mat(d, ω, qₓ, ϵ, μ, ξ, ζ)

Calculate the $M$ matrix according to Eq. 2.52 in [1].

Arguments

d: Thickness $[m]$.

ω: Angular frequency $[\frac{1}{s}]$.

qₓ: Normalized in-plane component of the wavevector.

ϵ: Permittivity tensor.

μ: Permeability tensor.

ξ, ζ: Optical rotation tensors.

Examples

julia> GeneralizedTransferMatrixMethod.M_mat(1e-6, 1.8e14, 0.3, 3ones(3,3), ones(3,3), zeros(3,3), zeros(3,3))
4×4 StaticArraysCore.SMatrix{4, 4, ComplexF64, 16} with indices SOneTo(4)×SOneTo(4):
 0.983821+0.179152im          0.0-0.0im          …       0.0-0.0im
      0.0-0.0im          0.983821+0.179152im             0.0-0.0im
      0.0-0.0im       3.22921e-11-1.77333e-10im          0.0-0.0im
      0.0-0.0im               0.0-0.0im             0.983821+0.179152im

References

GeneralizedTransferMatrixMethod.K_mat — Function

K_mat(n, cos_θ)

Calculate the $K$ matrix according to Eq. 3.16 in [1].

Arguments

n: Refractive index.

cos_θ: Cosine of angle between the k-vector and the surface normal.

Examples

julia> GeneralizedTransferMatrixMethod.K_mat(1.8, cos(π/4))
4×4 StaticArraysCore.SMatrix{4, 4, Float64, 16} with indices SOneTo(4)×SOneTo(4):
 0.0          0.707107     0.0         -0.707107
 1.0          0.0          1.0          0.0
 0.00337852   0.0         -0.00337852   0.0
 0.0         -0.00477795   0.0         -0.00477795

References

GeneralizedTransferMatrixMethod.calculate_layer_M — Function

calculate_layer_M(λ, qₓ, layer)

Calculate the $M$ matrix for a single layer.

Arguments

λ: Wavelength $[m]$.

qₓ: Normalized in-plane component of the wavevector.

layer: Material Layer.

Examples

julia> GeneralizedTransferMatrixMethod.calculate_layer_M(12.3e-6, 0.5, Layer(d = 2e-6))
4×4 StaticArraysCore.SMatrix{4, 4, ComplexF64, 16} with indices SOneTo(4)×SOneTo(4):
 0.63346+0.0im             0.0+0.0im         …      0.0-252.451im
     0.0+0.0im         0.63346+0.0im                0.0+0.0im
     0.0+0.0im             0.0+0.00177875im         0.0+0.0im
     0.0-0.00237167im      0.0+0.0im            0.63346+0.0im

GeneralizedTransferMatrixMethod.calculate_structure_M — Function

calculate_structure_M(λ, α, strct)

Calculate the $M$ matrix for a layered structure.

Arguments

λ: Wavelength $[m]$.

α: Angle of incidence.

strct: Structure of type LayeredStructure.

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate=Au());

julia> GeneralizedTransferMatrixMethod.calculate_structure_M(12.3e-6, deg2rad(20), S)
4×4 StaticArraysCore.SMatrix{4, 4, ComplexF64, 16} with indices SOneTo(4)×SOneTo(4):
 0.501778-0.00554054im       0.0+0.0im         …       0.0+0.0im
     -0.0+0.0im         0.471735-0.00589861im     -0.46795-0.0058937im
 0.498222+0.00554054im      -0.0+0.0im                -0.0+0.0im
     -0.0+0.0im         -0.46795-0.0058937im      0.471735-0.00589861im

Helper functions

GeneralizedTransferMatrixMethod.convert_to_wavelength — Function

convert_to_wavelength(x)

Function to dispatch on for extensions that use other wavelength/frequency/energy units. The default implementation is just the identity. Any function that has a wavelength argument uses this function.

GeneralizedTransferMatrixMethod.nan_to_zero — Function

nan_to_zero(x)

Return 0 if the input is NaN, othewise return x.

GeneralizedTransferMatrixMethod.build_dir — Function

build_dir(f)

Creats path to files in build directory.

GeneralizedTransferMatrixMethod.euler_mat — Function

euler_mat(θ, ϕ, ψ)

Calculate the Euler matrix from the Euler angles θ, ϕ, and ψ.

    euler_mat(layer::Layer)

Calculate the Euler matrix for a Layer.

GeneralizedTransferMatrixMethod.basis_change — Function

basis_change(M,T)

Change basis of matrix M using the basis transformation matrix T.

GeneralizedTransferMatrixMethod.T_mat — Constant

Basis tarnsformation matrix to change from a linear polarization basis to a circular polarization basis.

GeneralizedTransferMatrixMethod.circ_coeffs — Function

circ_coeffs(cₚₚ, cₛₛ, cₚₛ, cₛₚ)

Transform transmission/reflection coefficients from a linear to a circula basis. The output has the form (c_RR, c_LL, c_RL, c_LR).

Examples

julia> GeneralizedTransferMatrixMethod.circ_coeffs(0.23, 0.24, 0.14, 0.14)
(0.235 + 0.0im, 0.235 + 0.0im, -0.0050000000000000044 + 0.14im, -0.0050000000000000044 - 0.14im)

GeneralizedTransferMatrixMethod.basis_selector — Function

basis_selector(coeffs, basis)

Return linear input coeffs in given basis, either :linear or :circular.

GeneralizedTransferMatrixMethod.calculate_cos_αt — Function

calculate_cos_αt(λ, α, strct)

Calculate the cosine outgoing angle for a layered structure strct and angle of incidence α. Returns the tuple (cos_αt, n_inc, n_out, qₓ), where cos_αt is the cosine of the outgoing angle, n_inc is the refractive index of the incident medium, n_out is the refractive index of the outgoing medium, and qₓ is the normalized in-plane wavevector component.

We make sure to choose the correct branch when calculating the cosine for more see Section 5 in [1].

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate = SiC());

julia> GeneralizedTransferMatrixMethod.calculate_cos_αt(12e-6, deg2rad(14), S)
(1.0010924505072016 + 6.976717567528191e-5im, 1.0 + 0.0im, 0.1648892944698832 + 5.166277293025589im, 0.24192189559966773 + 0.0im)

References

Permittivities

GeneralizedTransferMatrixMethod.lorentz_osc — Function

lorentz_osc(f, fₗₒ, fₜₒ, γ)

Calculate a single lorentz oscillator (without ϵ∞). The definition is taken from [1].

Arguments

f: Frequency.

fₗₒ: Frequency of the longitudinal optical phonon.

fₜₒ: Frequency of the transverse optical phonon.

γ: Damping factor of the Lorentzian line shape.

References

GeneralizedTransferMatrixMethod.ϵ_drude — Function

ϵ_drude(f, fₚ, γ)

Calculate the permitivity from Drude model.

Arguments

f: Frequency.

fₚ: Plasma frequency.

γ: Mean collision rate.

GeneralizedTransferMatrixMethod.parse_permittivity — Function

parse_permittivity(func)

Function used inside the @permittivity macro to dispatch on for extensions. The default implementation just returns the original function.

GeneralizedTransferMatrixMethod.ϵ_x_MoO₃ — Function

ϵ_x_MoO₃(λ)

Calculate the x principal component of the permitivity tensor of MoO₃. The parameters are taken from [1].

Arguments

λ: Wavelength [m].

References

GeneralizedTransferMatrixMethod.ϵ_y_MoO₃ — Function

ϵ_y_MoO₃(λ)

Calculate the y principal component of the permitivity tensor of MoO₃. The parameters are taken from [1].

Arguments

λ: Wavelength [m].

References

GeneralizedTransferMatrixMethod.ϵ_z_MoO₃ — Function

ϵ_z_MoO₃(λ)

Calculate the z principal component of the permitivity tensor of MoO₃. The parameters are taken from [1].

Arguments

λ: Wavelength [m].

References

1Mackay, T. G. & Lakhtakia, A. The Transfer-Matrix Method in Electromagnetics and Optics. vol. 1 (2020).

1Mackay, T. G. & Lakhtakia, A. The Transfer-Matrix Method in Electromagnetics and Optics. vol. 1 (2020).

1Mackay, T. G. & Lakhtakia, A. The Transfer-Matrix Method in Electromagnetics and Optics. vol. 1 (2020).

1Mackay, T. G. & Lakhtakia, A. The Transfer-Matrix Method in Electromagnetics and Optics. vol. 1 (2020).

1Mackay, T. G. & Lakhtakia, A. The Transfer-Matrix Method in Electromagnetics and Optics. vol. 1 (2020).

1Byrnes, S. J. Multilayer Optical Calculations. arXiv:1603.02720 [Physics], December 30, 2020. http://arxiv.org/abs/1603.02720.

1Álvarez-Pérez, G. et al. Infrared Permittivity of the Biaxial van Der Waals Semiconductor α-MoO₃ from Near- and Far-Field Correlative Studies. Adv. Mater. 32, 1908176 (2020).

1Álvarez-Pérez, G. et al. Infrared Permittivity of the Biaxial van Der Waals Semiconductor α-MoO₃ from Near- and Far-Field Correlative Studies. Adv. Mater. 32, 1908176 (2020).

1Álvarez-Pérez, G. et al. Infrared Permittivity of the Biaxial van Der Waals Semiconductor α-MoO₃ from Near- and Far-Field Correlative Studies. Adv. Mater. 32, 1908176 (2020).

1Álvarez-Pérez, G. et al. Infrared Permittivity of the Biaxial van Der Waals Semiconductor α-MoO₃ from Near- and Far-Field Correlative Studies. Adv. Mater. 32, 1908176 (2020).