Public

Exported macros

GeneralizedTransferMatrixMethod.@permittivity — Macro

permmittivity(name, func)

This macro takes the name of a material and the permittivity func, depending on λ and returning the permittivity matrix. It generates the permittivity funcion ϵ_Name and the Layer function Name.

Note

The first letter of the name is always capitalised.

Exported functions

GeneralizedTransferMatrixMethod.calculate_layer_properties — Function

calculate_layer_properties(layer::Layer, ζ::Real, λ::Real)

Calculate the properties of a single layer in a layered structure and returns a struct of type LayerProperties.

Input

layer – Layer.

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

λ – Wavelength [m].

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

GeneralizedTransferMatrixMethod.calculate_structure_properties — Function

calculate_structure_properties(ζ::Real, λ::Real, strct::LayeredStructure)

Calculate the properties of a layered structure and returns a struct of type StructureProperties.

Input

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

λ – Wavelength [m].

strct – Layered structure.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

calculate_structure_properties(ζ::Real, λ::Unitful.Length,
                               strct::LayeredStructure)

Calculate the properties of a layered structure using Unitful quantities and returns a struct of type StructureProperties.

Input

ζ – In-plane reduced wavevector kₓ/k₀ in the system.

λ – Wavelength.

strct – Layered structure.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

GeneralizedTransferMatrixMethod.reflection — Function

reflection(r)

Calculate the reflection from the reflection coefficient. (Just square moduli)

Input

r – Reflection coefficient.

reflection(props::StructureProperties)

Calculate the reflection from a StructureProperties type. Returns the reflection for different polarisations in the order Rₚₚ, Rₛₛ, Rₚₛ, Rₛₚ.

Input

props – Layered material reperesented by a StructureProperties type.

GeneralizedTransferMatrixMethod.transmission — Function

transmission(ζ, props::StructureProperties)

Calculate the total transmission from a StructureProperties type. Returns the transmission in the order Tₚ, Tₛ, where the indices indicate the incoming polarisation.

Input

ζ – In-plane reduced wavevector kₓ/k₀ of the system.

props – Layered material reperesented by a StructureProperties type.

GeneralizedTransferMatrixMethod.reflection_coeffs — Function

reflection_coeffs(Γ)

Calculate the reflection coefficients from the full transfer matrix Γ (Γ* in the reference). Returns the reflection coefficients in the order rₚₚ, rₛₛ, rₚₛ, rₛₚ.

Input

Γ – Full transfer matrix in Yeh's formalism.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

reflection_coeffs(props::StructureProperties)

Calculate the reflection coefficients from a StructureProperties type. Returns the reflection coefficients for different polarisations in the order rₚₚ, rₛₛ, rₚₛ, rₛₚ.

Input

props – Layered material reperesented by a StructureProperties type.

GeneralizedTransferMatrixMethod.transmission_coeffs — Function

transmission_coeffs(Γ)

Calculate the transmission coefficients from the full transfer matrix Γ (Γ* in the reference). Returns the reflection coefficients in the order tₚₚ, tₛₛ, tₚₛ, tₛₚ.

Input

Γ – Full transfer matrix in Yeh's formalism.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

transmission_coeffs(props::StructureProperties)

Calculate the transmission coefficients from a StructureProperties type. Returns the transmission coefficients for different polarisations in the order tₚₚ, tₛₛ, tₚₛ, tₛₚ.

Input

props – Layered material reperesented by a StructureProperties type.

Exported types

GeneralizedTransferMatrixMethod.Layer — Type

struct Layer
    ϵ::Function
    d::Real
    θ::Real
    ϕ::Real
    ψ::Real
end

Type that represents a layer in a layered structur. Boundary layers have d = 0.

Fields

ϵ – Permitivity tensor in diagonal form (default: ϵ_vacuum).

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

Reference

The definition of the Euler matrix is taken from Optical Waves in Layered Media by Pochi Yeh.

GeneralizedTransferMatrixMethod.LayeredStructure — Type

struct LayeredStructure
    superstrate::Layer
    layers::AbstractVector{Layer}
    substrate::Layer
end

Type that represents a layered structure.

Fields

superstrate – Superstrate of the structure.

layers – List of layers of the structure starting with the layer underneath the superstrate and ending with the layer above the substrate (default = []).

substrate – Subststrate of the structure.

GeneralizedTransferMatrixMethod.LayerProperties — Type

struct LayerProperties
    T::AbstractMatrix
    P::AbstractMatrix
    A::AbstractMatrix
    γ::AbstractMatrix
    q::AbstractVector
    Ψ::AbstractVector
    S::AbstractVector
    Δ::AbstractMatrix
    a::AbstractMatrix
    M::AbstractMatrix
end

Type that holds all properties of a single layer in a structure. For boundary layers P and T are identity matrices.

The properties of this type only make sense in combination with the wavelength λ and the normalized in-plane wavevector component ζ.

Fields

T – Transmission matrix of the layer.

P – Progation matrix of the layer.

A – Dynamical matrix of the layer.

γ – Normalized γ vectors in matrix form.

q – Four eigenvalues (z-components of the wavevector) of the layer.

Ψ – Four eigenmodes of the layer.

S – Poynting vectors of the four eigenmodes of the layer.

Δ – Δ matrix.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

GeneralizedTransferMatrixMethod.StructureProperties — Type

struct StructureProperties
            Γ::AbstractMatrix
            superstrate::LayerProperties
            layers::AbstractVector{LayerProperties}
            substrate::LayerProperties
    end

Type that holds all properties of an entire layered structure.

The properties of this type only make sense in combination with the wavelength λ and the normalized in-plane wavevector component ζ.

Fields

Γ – Full transfer matrix Γ* (see reference).

superstrate – Properties of the superstrate layer.

layers – List of properties of the intermediate layers.

substrate – Properties of the substrate layer.

Reference

The definition is taken from Passler and Paarmann 2017 and Passler and Paarmann 2019 (erratum).

Exported permittivities

GeneralizedTransferMatrixMethod.ϵ_vacuum — Function

ϵ_vacuum(λ)

Calculate the permittivity of vacuum. Return thes identity matrix.

Input

λ – Wavelengeth [m].

GeneralizedTransferMatrixMethod.ϵ_Ag — Function

ϵ_Ag(λ)

Calculate permittivity tensor of Ag.

Input

λ – Wavelength [m].

ϵ_Ag(λ::Unitful.Length)

Calculate permittivity tensor of Ag using Unitful quantities.

Input

λ – Wavelength.

GeneralizedTransferMatrixMethod.ϵ_Au — Function

ϵ_Au(λ)

Calculate permittivity tensor of Au.

Input

λ – Wavelength [m].

ϵ_Au(λ::Unitful.Length)

Calculate permittivity tensor of Au using Unitful quantities.

Input

λ – Wavelength.

GeneralizedTransferMatrixMethod.ϵ_SiC — Function

ϵ_SiC(λ)

Calculate permittivity tensor of SiC.

Input

λ – Wavelength [m].

ϵ_SiC(λ::Unitful.Length)

Calculate permittivity tensor of SiC using Unitful quantities.

Input

λ – Wavelength.

GeneralizedTransferMatrixMethod.ϵ_MoO₃ — Function

ϵ_MoO₃(λ)

Calculate permittivity tensor of MoO₃.

Input

λ – Wavelength [m].

ϵ_MoO₃(λ::Unitful.Length)

Calculate permittivity tensor of MoO₃ using Unitful quantities.

Input

λ – Wavelength.

Exported material layers

GeneralizedTransferMatrixMethod.Ag — Function

Ag(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of Ag using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

Ag(d, θ, ϕ, ψ)

Define a Layer of Ag.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

Ag(d::Unitful.Length, θ, ϕ, ψ)

Define a Layer of Ag using Unitful quantities.

Input

d – Thickness of the layer.

θ – θ Euler angle.

ϕ – ϕ Euler angle.

ψ – ψ Euler angle.

GeneralizedTransferMatrixMethod.Au — Function

Au(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of Au using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

Au(d, θ, ϕ, ψ)

Define a Layer of Au.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

Au(d::Unitful.Length, θ, ϕ, ψ)

Define a Layer of Au using Unitful quantities.

Input

d – Thickness of the layer.

θ – θ Euler angle.

ϕ – ϕ Euler angle.

ψ – ψ Euler angle.

GeneralizedTransferMatrixMethod.SiC — Function

SiC(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of SiC using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

SiC(d, θ, ϕ, ψ)

Define a Layer of SiC.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

SiC(d::Unitful.Length, θ, ϕ, ψ)

Define a Layer of SiC using Unitful quantities.

Input

d – Thickness of the layer.

θ – θ Euler angle.

ϕ – ϕ Euler angle.

ψ – ψ Euler angle.

GeneralizedTransferMatrixMethod.MoO₃ — Function

MoO₃(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of MoO₃ using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

MoO₃(d, θ, ϕ, ψ)

Define a Layer of MoO₃.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

MoO₃(d::Unitful.Length, θ, ϕ, ψ)

Define a Layer of MoO₃ using Unitful quantities.

Input

d – Thickness of the layer.

θ – θ Euler angle.

ϕ – ϕ Euler angle.

ψ – ψ Euler angle.