Public

Exported macros

GeneralizedTransferMatrixMethod.@permittivity — Macro

permmittivity(name, func)

This macro takes the name of a material and the permittivity func, depending on λ and returning the permittivity matrix. It generates the permittivity funcion ϵ_Name and the Layer function Name.

Note

The first letter of the name is always capitalised.

Exported functions

GeneralizedTransferMatrixMethod.reflection_coeffs — Function

reflection_coeffs(λ, α, strct; basis=:linear)

Calculate the reflection coefficients.

Arguments

λ: Wavelength $[m]$.

α: Angle of incidence.

strct: Structure of type LayeredStructure.

basis=:linear: Returns (rₚₚ, rₛₛ, rₚₛ, rₛₚ) in :linear basis and (r_RR, r_LL, r_RL, r_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate=Au());

julia> reflection_coeffs(12.3e-6, deg2rad(20), S)
(0.9916655707450739 + 0.02489351332010354im, -0.9926693143906693 - 0.02200268526284087im, 0.0 + 0.0im, -0.0 + 0.0im)

reflection_coeffs(M; basis=:linear)

Calculate reflection coefficients from M matrix. Returns (rₚₚ, rₛₛ, rₚₛ, rₛₚ) in :linear basis and (r_RR, r_LL, r_RL, r_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate=Au());

julia> M = GeneralizedTransferMatrixMethod.calculate_structure_M(12.3e-6, deg2rad(20), S);

julia> reflection_coeffs(M)
(0.9916655707450739 + 0.02489351332010354im, -0.9926693143906693 - 0.02200268526284087im, 0.0 + 0.0im, -0.0 + 0.0im)

GeneralizedTransferMatrixMethod.transmission_coeffs — Function

transmission_coeffs(λ, α, strct; basis=:linear)

Calculate the transmission coefficients.

Arguments

λ: Wavelength $[m]$.

α: Angle of incidence.

strct: Structure of type LayeredStructure.

basis=:linear: Returns (tₚₚ, tₛₛ, tₚₛ, tₛₚ) in :linear basis and (t_RR, t_LL, t_RL, t_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate = SiC());

julia> transmission_coeffs(12e-6, deg2rad(18), S)
(0.08647336439281672 - 0.3667267570954972im, 0.07570669303896788 - 0.35120281878473003im, 0.0 + 0.0im, 0.0 + 0.0im)

transmission_coeffs(M; basis=:linear)

Calculate transmission coefficients from M matrix. Returns (tₚₚ, tₛₛ, tₚₛ, tₛₚ) in :linear basis and (t_RR, t_LL, t_RL, t_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), substrate = SiC());

julia> M = GeneralizedTransferMatrixMethod.calculate_structure_M(12e-6, deg2rad(18), S);

julia> transmission_coeffs(M)
(0.08647336439281672 - 0.3667267570954972im, 0.07570669303896788 - 0.35120281878473003im, 0.0 + 0.0im, 0.0 + 0.0im)

GeneralizedTransferMatrixMethod.calculate_reflection — Function

calculate_reflection(λ, α, strct; basis=:linear)

Calculate the reflection of a layered structure.

Arguments

λ: Wavelength $[m]$.

α: Angle of incidence.

strct: Structure of type LayeredStructure.

basis=:linear: Returns (Rₚₚ, Rₛₛ, Rₚₛ, Rₛₚ) in :linear basis and (R_RR, R_LL, R_RL, R_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), layers=[MoO₃(d = 1e-6)], substrate = Layer());

julia> calculate_reflection(12.5e-6, deg2rad(23), S)
(0.19265761641397677, 0.31745983790535026, 0.0, 0.0)

GeneralizedTransferMatrixMethod.calculate_transmission — Function

calculate_transmission(λ, α, strct; basis=:linear)

Calculate the transmission of a layered structure.

Arguments

λ: Wavelength $[m]$.

α: Angle of incidence.

strct: Structure of type LayeredStructure.

basis=:linear: Returns (Tₚₚ, Tₛₛ, Tₚₛ, Tₛₚ) in :linear basis and (T_RR, T_LL, T_RL, T_LR) in :circular basis.

Examples

julia> S = LayeredStructure(superstrate=Layer(), layers=[MoO₃(d = 1e-6)], substrate = Layer());

julia> calculate_transmission(12.5e-6, deg2rad(23), S)
(0.31887703646260473, 0.6139558802434731, 0.0, 0.0)

Exported types

GeneralizedTransferMatrixMethod.Layer — Type

struct Layer
    ϵ::Function = ϵ_vacuum
    μ::Function = μ_vacuum
    ξ::Function = ξ_vacuum
    ζ::Function = ζ_vacuum
    d::Real = 0
    θ::Real = 0
    ϕ::Real = 0
    ψ::Real = 0
end

Type that represents a layer in a layered structur. Boundary layers have d = 0.

Arguments

ϵ: Relative permitivity tensor (default: ϵ_vacuum).

μ: Relative permeability tensor (default: μ_vacuum).

ξ, ζ: Optical rotation tensors (default: ξ_vacuum, ζ_vacuum).

d: Thickness of the layer [m] (default: 0).

θ: θ Euler angle [rad] (default: 0).

ϕ: ϕ Euler angle [rad] (default: 0).

ψ: ψ Euler angle [rad] (default: 0).

The definition of the Euler angles is taken from [1].

References

GeneralizedTransferMatrixMethod.LayeredStructure — Type

struct LayeredStructure
    superstrate::Layer
    layers::AbstractVector{Layer} = []
    substrate::Layer
end

Type that represents a layered structure.

Arguments

superstrate: Superstrate of the structure.

layers: List of layers of the structure starting with the layer underneath the superstrate and ending with the layer above the substrate (default = []).

substrate: Subststrate of the structure.

Exported permittivities

GeneralizedTransferMatrixMethod.ϵ_vacuum — Function

ϵ_vacuum(λ)

Calculate the relative permittivity of vacuum. Always returns the identity matrix.

GeneralizedTransferMatrixMethod.μ_vacuum — Function

μ_vacuum(λ)

Calculate the relative permeability of vacuum. Always returns the identity matrix.

GeneralizedTransferMatrixMethod.ξ_vacuum — Function

Optical rotation tensor

GeneralizedTransferMatrixMethod.ζ_vacuum — Function

Optical rotation tensor

GeneralizedTransferMatrixMethod.ϵ_Ag — Function

ϵ_Ag(λ)

Calculate permittivity tensor of Ag.

Input

λ – Wavelength [m].

GeneralizedTransferMatrixMethod.ϵ_Au — Function

ϵ_Au(λ)

Calculate permittivity tensor of Au.

Input

λ – Wavelength [m].

GeneralizedTransferMatrixMethod.ϵ_SiC — Function

ϵ_SiC(λ)

Calculate permittivity tensor of SiC.

Input

λ – Wavelength [m].

GeneralizedTransferMatrixMethod.ϵ_MoO₃ — Function

ϵ_MoO₃(λ)

Calculate permittivity tensor of MoO₃.

Input

λ – Wavelength [m].

Exported material layers

GeneralizedTransferMatrixMethod.Ag — Function

Ag(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of Ag using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

Ag(d, θ, ϕ, ψ)

Define a Layer of Ag.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

GeneralizedTransferMatrixMethod.Au — Function

Au(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of Au using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

Au(d, θ, ϕ, ψ)

Define a Layer of Au.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

GeneralizedTransferMatrixMethod.SiC — Function

SiC(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of SiC using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

SiC(d, θ, ϕ, ψ)

Define a Layer of SiC.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

GeneralizedTransferMatrixMethod.MoO₃ — Function

MoO₃(;d = 0, θ = 0, ϕ = 0, ψ = 0)

Define a Layer of MoO₃ using keyword arguments.

Input

Suitable quantities with units from Unitful also work.

d – Thickness of the layer [m] (default: 0).

θ – θ Euler angle [rad] (default: 0).

ϕ – ϕ Euler angle [rad] (default: 0).

ψ – ψ Euler angle [rad] (default: 0).

MoO₃(d, θ, ϕ, ψ)

Define a Layer of MoO₃.

Input

d – Thickness of the layer [m].

θ – θ Euler angle [rad].

ϕ – ϕ Euler angle [rad].

ψ – ψ Euler angle [rad].

1Yeh, P. Optical Waves in Layered Media. Wiley Series in Pure and Applied Optics. Wiley, 2005.